backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==10.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
python-multipart==0.0.6
textstat==0.7.3
nltk==3.8.1
google-generativeai==0.7.0
python-dotenv==1.0.1
gunicorn==21.2.0 